+++ test/factories/exam_version.rb
@@ -7,21 +7,21 @@ FactoryBot.define do
     trait :cs2500_v1 do
       name { 'CS2500 Midterm Version 1' }
       transient do
-        upload { fixture_zip 'cs2500midterm-v1' }
+        upload { create(:upload, :cs2500_v1) }
       end
     end
 
     trait :cs2500_v2 do
       name { 'CS2500 Midterm Version 2' }
       transient do
-        upload { fixture_zip 'cs2500midterm-v2' }
+        upload { create(:upload, :cs2500_v2) }
       end
     end
 
     trait :cs3500_v1 do
       name { 'CS3500 Final Version 1' }
       transient do
-        upload { fixture_zip 'cs3500final' }
+        upload { create(:upload, :cs3500_v1) }
       end
     end
